+++ country_generator/lib/src/extension.dart
@@ -82,6 +82,7 @@ extension CountryClassStringExtension on Country {
       esmMember: $esmMember,
       altCurrency: ${altCurrency == null ? null : '\'${altCurrency!.replaceAll('\'', '\\\'')}\''},
       isoShortNameByLocale: ${json.encode(isoShortNameByLocale)},
+      subdivision: ${subdivision.toClassString()},
       )''';
   }
 }
@@ -236,5 +237,37 @@ extension WorldRegionClassStringExtension on WorldRegion {
       default:
         throw ArgumentError('Unsupported wireName: $wireName');
     }
+  }
+}
+
+extension SubdivisionGeoDataClassStringExtension on SubdivisionGeoData {
+  String toClassString() {
+    return '''SubdivisionGeoData(
+          latitude: $latitude,
+          longitude: $longitude,
+          maxLatitude: $maxLatitude,
+          maxLongitude: $maxLongitude,
+          minLatitude: $minLatitude,
+          minLongitude: $minLongitude,
+          )''';
+  }
+}
+
+extension SubdivisionClassStringExtension on Subdivision {
+  String toClassString() {
+    return '''Subdivision(
+              name: '${name.replaceAll('\'', '\\\'')}',
+              code: '$code',
+              geo: ${geo == null ? null : geo!.toClassString()},
+              translation: ${json.encode(translation)},
+              type: '$type',
+              )''';
+  }
+}
+
+/// Extension class for generating class string with [List<int>]
+extension ListSubdivisionExtension on List<Subdivision> {
+  String toClassString() {
+    return '[${map((e) => e.toClassString()).join(',')}]';
   }
 }