+++ country/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 4.0.0
+
+* added subdivision to country object
+* updated submodule
+* updated generated country with new data
+
+## 3.0.1+1
+
+* update package topics
+
 ## 3.0.1
 
 * fix dart sdk constraint