+++ flutter_country/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.0.0
+
+* upgrade country to '4.0.0'
+
+## 3.0.0+1
+
+* update package topics
+
 ## 3.0.0
 
 * upgrade country to '3.0.0'